add blueprint and sales routes
